convert py3 qt5 by script
