+++ AnkiHabitica_v2.1/AnkiHabitica/logging/__init__.py
@@ -23,7 +23,7 @@ Copyright (C) 2001-2014 Vinay Sajip. All Rights Reserved.
 To use, simply 'import logging' and log away!
 """
 
-import sys, os, time, cStringIO, traceback, warnings, weakref, collections
+import sys, os, time, io, traceback, warnings, weakref, collections
 
 __all__ = ['BASIC_FORMAT', 'BufferingFormatter', 'CRITICAL', 'DEBUG', 'ERROR',
            'FATAL', 'FileHandler', 'Filter', 'Formatter', 'Handler', 'INFO',
@@ -39,7 +39,7 @@ except ImportError:
     codecs = None
 
 try:
-    import thread
+    import _thread
     import threading
 except ImportError:
     thread = None
@@ -54,7 +54,7 @@ __date__    = "07 February 2010"
 #   Miscellaneous module data
 #---------------------------------------------------------------------------
 try:
-    unicode
+    str
     _unicode = True
 except NameError:
     _unicode = False
@@ -175,7 +175,7 @@ def addLevelName(level, levelName):
         _releaseLock()
 
 def _checkLevel(level):
-    if isinstance(level, (int, long)):
+    if isinstance(level, int):
         rv = level
     elif str(level) == level:
         if level not in _levelNames:
@@ -278,10 +278,10 @@ class LogRecord(object):
         self.lineno = lineno
         self.funcName = func
         self.created = ct
-        self.msecs = (ct - long(ct)) * 1000
+        self.msecs = (ct - int(ct)) * 1000
         self.relativeCreated = (self.created - _startTime) * 1000
         if logThreads and thread:
-            self.thread = thread.get_ident()
+            self.thread = _thread.get_ident()
             self.threadName = threading.current_thread().name
         else:
             self.thread = None
@@ -298,7 +298,7 @@ class LogRecord(object):
                 # for an example
                 try:
                     self.processName = mp.current_process().name
-                except StandardError:
+                except Exception:
                     pass
         if logProcesses and hasattr(os, 'getpid'):
             self.process = os.getpid()
@@ -320,7 +320,7 @@ class LogRecord(object):
             msg = str(self.msg)
         else:
             msg = self.msg
-            if not isinstance(msg, basestring):
+            if not isinstance(msg, str):
                 try:
                     msg = str(self.msg)
                 except UnicodeError:
@@ -435,7 +435,7 @@ class Formatter(object):
         This default implementation just uses
         traceback.print_exception()
         """
-        sio = cStringIO.StringIO()
+        sio = io.StringIO()
         traceback.print_exception(ei[0], ei[1], ei[2], None, sio)
         s = sio.getvalue()
         sio.close()
@@ -865,9 +865,9 @@ class StreamHandler(Handler):
                 stream.write(fs % msg)
             else:
                 try:
-                    if (isinstance(msg, unicode) and
+                    if (isinstance(msg, str) and
                         getattr(stream, 'encoding', None)):
-                        ufs = u'%s\n'
+                        ufs = '%s\n'
                         try:
                             stream.write(ufs % msg)
                         except UnicodeEncodeError:
@@ -1034,9 +1034,9 @@ class Manager(object):
         placeholder to now point to the logger.
         """
         rv = None
-        if not isinstance(name, basestring):
+        if not isinstance(name, str):
             raise TypeError('A logger name must be string or Unicode')
-        if isinstance(name, unicode):
+        if isinstance(name, str):
             name = name.encode('utf-8')
         _acquireLock()
         try:
@@ -1099,7 +1099,7 @@ class Manager(object):
         """
         name = alogger.name
         namelen = len(name)
-        for c in ph.loggerMap.keys():
+        for c in list(ph.loggerMap.keys()):
             #The if means ... if not c.parent.name.startswith(nm)
             if c.parent.name[:namelen] != name:
                 alogger.parent = c.parent
